moved the route generator to routes.js
